extract addCardToInventoryReceiving
